consolidate service and calculator models
