+++ cmd/jiracollector/jiracollector/jiraissuecollector.go
@@ -10,17 +10,17 @@ import (
 	"github.com/stebennett/squad-dashboard/cmd/jiracollector/models"
 	jiramodels "github.com/stebennett/squad-dashboard/pkg/jira/models"
 	"github.com/stebennett/squad-dashboard/pkg/jira/repo/issuerepository"
-	"github.com/stebennett/squad-dashboard/pkg/jiraservice"
+	"github.com/stebennett/squad-dashboard/pkg/jira/service"
 	"github.com/stebennett/squad-dashboard/pkg/paginator"
 )
 
 type JiraIssueCollector struct {
 	repo      issuerepository.IssueRepository
-	jira      *jiraservice.JiraService
+	jira      *service.JiraService
 	epicField string
 }
 
-func NewJiraIssueCollector(jira *jiraservice.JiraService, repo issuerepository.IssueRepository, epicField string) *JiraIssueCollector {
+func NewJiraIssueCollector(jira *service.JiraService, repo issuerepository.IssueRepository, epicField string) *JiraIssueCollector {
 	return &JiraIssueCollector{
 		repo:      repo,
 		jira:      jira,
@@ -37,7 +37,7 @@ func (jc *JiraIssueCollector) execute(project string, startAt int, maxResults in
 	saveTransition func(ctx context.Context, issueKey string, jiraTransitions []jiramodels.JiraTransition) (int64, error),
 	saveIssueLabels func(ctx context.Context, issueKey string, labels []string) (int64, error)) error {
 
-	query := jiraservice.JiraSearchQuery{
+	query := service.JiraSearchQuery{
 		Jql:        jql,
 		Fields:     []string{"summary", "issuetype", epicField, "created", "updated", "labels"},
 		Expand:     []string{"changelog"},
