--- pkg/calculatormodels/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "calculatormodels",
-    srcs = ["calculatormodels.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/calculatormodels",
-    visibility = ["//visibility:public"],
-    deps = ["//vendor/github.com/lib/pq"],
-)
