+++ pkg/jira/repo/issuerepository/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
     importpath = "github.com/stebennett/squad-dashboard/pkg/jira/repo/issuerepository",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/calculatormodels",
         "//pkg/jira/models",
         "//vendor/github.com/lib/pq",
     ],
