+++ pkg/jira/service/BUILD.bazel
@@ -2,7 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "jiraservice",
-    srcs = ["jiraservice.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/jiraservice",
+    srcs = ["service.go"],
+    importpath = "github.com/stebennett/squad-dashboard/pkg/jira/service",
     visibility = ["//visibility:public"],
 )
